Enhancing structure of project
