+++ Stroke_Risk_Prediction/templates/risk.html
@@ -5,41 +5,9 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Stroke Risk Prediction</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            text-align: center;
-            padding: 20px;
-        }
-        form {
-            margin: auto;
-            width: 50%;
-            padding: 10px;
-            border: 1px solid #ccc;
-            border-radius: 10px;
-            background-color: #f9f9f9;
-        }
-        label {
-            display: block;
-            margin-top: 10px;
-        }
-        input {
-            width: calc(100% - 16px);
-            padding: 8px;
-            margin-top: 5px;
-            margin-bottom: 10px;
-        }
-        button {
-            padding: 10px 15px;
-            background-color: #007BFF;
-            color: white;
-            border: none;
-            border-radius: 5px;
-        }
-        button:hover {
-            background-color: #0056b3;
-        }
-    </style>
+    <link rel="stylesheet" href="{{ url_for('stroke.static', filename='style.css') }}">
+
+   
 </head>
 
 <body>
@@ -49,7 +17,7 @@
 
     <div class="container-fluid">
         <div class="col-lg-12">
-            <form action="/predict" method="POST">
+            <form action="stroke-risk/predict" method="POST">
                 <br>
                 <br>
                 <div class="form-group">
