--- main/templates/diabetes.html
@@ -1,85 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Diabetes Prediction</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            text-align: center;
-            padding: 20px;
-        }
-        form {
-            margin: auto;
-            width: 50%;
-            padding: 10px;
-            border: 1px solid #ccc;
-            border-radius: 10px;
-            background-color: #f9f9f9;
-        }
-        label {
-            display: block;
-            margin-top: 10px;
-        }
-        input {
-            width: calc(100% - 16px);
-            padding: 8px;
-            margin-top: 5px;
-            margin-bottom: 10px;
-        }
-        button {
-            padding: 10px 15px;
-            background-color: #007BFF;
-            color: white;
-            border: none;
-            border-radius: 5px;
-        }
-        button:hover {
-            background-color: #0056b3;
-        }
-    </style>
-</head>
-<body>
-    <div class="container">
-        <h1>Diabetes Prediction</h1>
-        <form action="/predict" method="POST">
-            <div class="form-grid">
-                <div>
-                    <label for="Pregnancies">Pregnancies</label>
-                    <input type="number" step="any" name="Pregnancies" required>
-                </div>
-                <div>
-                    <label for="Glucose">Glucose</label>
-                    <input type="number" step="any" name="Glucose" required>
-                </div>
-                <div>
-                    <label for="BloodPressure">Blood Pressure</label>
-                    <input type="number" step="any" name="BloodPressure" required>
-                </div>
-                <div>
-                    <label for="SkinThickness">Skin Thickness</label>
-                    <input type="number" step="any" name="SkinThickness" required>
-                </div>
-                <div>
-                    <label for="Insulin">Insulin</label>
-                    <input type="number" step="any" name="Insulin" required>
-                </div>
-                <div>
-                    <label for="BMI">BMI</label>
-                    <input type="number" step="any" name="BMI" required>
-                </div>
-                <div>
-                    <label for="DiabetesPedigreeFunction">Diabetes Pedigree Function</label>
-                    <input type="number" step="any" name="DiabetesPedigreeFunction" required>
-                </div>
-                <div>
-                    <label for="Age">Age</label>
-                    <input type="number" step="any" name="Age" required>
-                </div>
-            </div>
-            <button type="submit" class="submit-btn">Predict</button>
-        </form>
-    </div>
-</body>
-</html>
